+++ src/ephys_link/bindings/mpm_binding.py
@@ -13,8 +13,8 @@ from typing import Any
 from requests import JSONDecodeError, get, put
 from vbl_aquarium.models.unity import Vector3, Vector4
 
-from ephys_link.util.base_binding import BaseBinding
-from ephys_link.util.common import scalar_mm_to_um, vector4_to_array
+from ephys_link.utils.base_binding import BaseBinding
+from ephys_link.utils.common import scalar_mm_to_um, vector4_to_array
 
 
 class MPMBinding(BaseBinding):
